requirements.txt
ipyvizzu==0.16.2
matplotlib==3.8.2
pandas==2.1.3
Pillow==10.1.0
Pillow==10.1.0
plotly==5.18.0
python_dateutil==2.8.2
st_pages==0.4.5
st_vizzu==0.1
streamlit==1.29.0
openpyxl==3.1.2
streamlit_extras==0.3.5
